Split the files align.* in many parts. Keep some of them together.
