+++ src/align_left_shift.h
@@ -1,0 +1,16 @@
+/**
+ * @file align_left_shift.h
+ *
+ * @author  Guy Maurel
+ * split from align.cpp
+ * @author  Ben Gardner
+ * @license GPL v2+
+ */
+
+#ifndef ALIGN_LEFT_SHIFT_H_INCLUDED
+#define ALIGN_LEFT_SHIFT_H_INCLUDED
+
+//! Align left shift operators '<<' (CT_ARITH?)
+void align_left_shift(void);
+
+#endif /* ALIGN_LEFT_SHIFT_H_INCLUDED */
