+++ src/align_oc_decl_colon.h
@@ -1,0 +1,21 @@
+/**
+ * @file align_oc_decl_colon.h
+ *
+ * @author  Guy Maurel
+ * split from align.cpp
+ * @author  Ben Gardner
+ * @license GPL v2+
+ */
+
+#ifndef ALIGN_OC_DECL_COLON_H_INCLUDED
+#define ALIGN_OC_DECL_COLON_H_INCLUDED
+
+
+/**
+ * Aligns OC declarations on the colon
+ * -(void) doSomething: (NSString*) param1
+ *                with: (NSString*) param2
+ */
+void align_oc_decl_colon(void);
+
+#endif /* ALIGN_OC_DECL_COLON_H_INCLUDED */
