+++ src/align_oc_msg_colons.h
@@ -1,0 +1,16 @@
+/**
+ * @file align_oc_msg_colons.h
+ *
+ * @author  Guy Maurel
+ * split from align.cpp
+ * @author  Ben Gardner
+ * @license GPL v2+
+ */
+
+#ifndef ALIGN_OC_MSG_COLONS_H_INCLUDED
+#define ALIGN_OC_MSG_COLONS_H_INCLUDED
+
+//! Aligns OC messages
+void align_oc_msg_colons(void);
+
+#endif /* ALIGN_OC_MSG_COLONS_H_INCLUDED */
