+++ src/align_oc_msg_spec.h
@@ -1,0 +1,18 @@
+/**
+ * @file align_oc_msg_spec.h
+ *
+ * @author  Guy Maurel
+ * split from align.cpp
+ * @author  Ben Gardner
+ * @license GPL v2+
+ */
+
+#ifndef ALIGN_OC_MSG_SPEC_H_INCLUDED
+#define ALIGN_OC_MSG_SPEC_H_INCLUDED
+
+#include "uncrustify_types.h"
+
+//! Aligns all function prototypes in the file.
+void align_oc_msg_spec(size_t span);
+
+#endif /* ALIGN_OC_MSG_SPEC_H_INCLUDED */
