+++ src/prototypes.h
@@ -54,25 +54,4 @@ static inline size_t next_tab_column(size_t col)
 }
 
 
-/**
- * Advances to the next tab stop if not currently on one.
- *
- * @param col  The current column
- * @return the next tabstop column
- */
-static inline size_t align_tab_column(size_t col)
-{
-   //if (col <= 0)
-   if (col == 0)
-   {
-      col = 1;
-   }
-   if ((col % uncrustify::options::output_tab_size()) != 1)
-   {
-      col = next_tab_column(col);
-   }
-   return(col);
-}
-
-
 #endif /* C_PARSE_PROTOTYPES_H_INCLUDED */
