+++ src/quick_align_again.h
@@ -1,0 +1,17 @@
+/**
+ * @file quick_align_again.h
+ *
+ * @author  Guy Maurel
+ * split from align.cpp
+ * @author  Ben Gardner
+ * @license GPL v2+
+ */
+
+#ifndef QUICK_ALIGN_AGAIN_H_INCLUDED
+#define QUICK_ALIGN_AGAIN_H_INCLUDED
+
+#include "uncrustify_types.h"
+
+void quick_align_again(void);
+
+#endif /* QUICK_ALIGN_AGAIN_H_INCLUDED */
